Use .env to store db password, add .env to .gitignore, add .idea to gitignore, separate models into separate files
