+++ parkhangs-expressapp/.gitignore
@@ -34,3 +34,6 @@ lerna-debug.log*
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+#.env file
+.env
