+++ parkhangs-expressapp/app.js
@@ -11,6 +11,11 @@ var app = express();
 var uuid = require('uuidv4');
 var cors = require('cors');
 
+//env file
+// require('dotenv').config();
+// //
+// // let PASSWORD = process.env.DB_PASSWORD;
+
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'jade');
