+++ parkhangs-expressapp/database/index.js
@@ -1,7 +1,11 @@
 const mongoose = require('mongoose');
+//import this statement to get the .env file
+require('dotenv').config();
+
+//this is how you get the value from the env file
+let PASSWORD = process.env.DB_PASSWORD;
 
 let DBNAME = "ParkHangs";
-let PASSWORD = "PUT DB PASSWORD HERE";
 
 let MONGODBCONNECTIONURL = `mongodb+srv://park_hangs_team:${PASSWORD}@parkhangs-ngrhm.mongodb.net/${DBNAME}?retryWrites=true&w=majority`;
